rename service and create installable package
