+++ nlp/nlp_service/resolver.py
