+++ nlp/pyproject.toml
@@ -1,0 +1,7 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "NLP"
+version = "0.0.1"
